+++ packages/scheduler/src/index.ts
@@ -1,0 +1,6 @@
+export { Scheduled, type ScheduledOptions } from './decorators/scheduled.js';
+export {
+  type ScheduledMethodMeta,
+  SchedulerService,
+} from './scheduler-service.js';
+export { createSchedulerPlugin } from './scheduler-transformer-plugin.js';